+++ resources/js/messages/Messages.vue
@@ -61,7 +61,7 @@
                             <button
                                 v-if="!received(msg)"
                                 class="text-secondary btn btn-link p-0 pb-1"
-                                v-b-tooltip.hover="deliveryStatus(msg)"
+                                v-b-tooltip.topleft :title="deliveryStatus(msg)"
                             >
                                 <svg class="ic ic-clock" v-if="!msg.seen_at"><use xlink:href="#icon-clock"></use></svg>
                                 <svg class="ic ic-check" v-else><use xlink:href="#icon-check"></use></svg>